# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: quiz
    description: Quiz question generation
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /quiz/questions:
    post:
      operationId: generateQuizQuestions
      tags: [quiz]
      summary: Generate quiz questions via AI
      description: Uses AI to generate unique quiz questions for a given category and difficulty
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GenerateQuestionsBody"
      responses:
        "200":
          description: Generated questions
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GeneratedQuestionsResponse"
        "500":
          description: Generation failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status
    GenerateQuestionsBody:
      type: object
      properties:
        category:
          type: string
          enum: [web, math, general]
        difficulty:
          type: string
          enum: [easy, medium, hard]
        count:
          type: integer
          minimum: 1
          maximum: 20
          default: 10
      required:
        - category
        - difficulty
    GeneratedQuestion:
      type: object
      properties:
        id:
          type: string
        question:
          type: string
        correctAnswer:
          type: string
        incorrectAnswers:
          type: array
          items:
            type: string
          minItems: 3
          maxItems: 3
        category:
          type: string
        difficulty:
          type: string
      required:
        - id
        - question
        - correctAnswer
        - incorrectAnswers
        - category
        - difficulty
    GeneratedQuestionsResponse:
      type: object
      properties:
        questions:
          type: array
          items:
            $ref: "#/components/schemas/GeneratedQuestion"
        source:
          type: string
          enum: [ai, fallback]
      required:
        - questions
        - source
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error