# lib/api-zod/src/generated/types/generateQuestionsBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { GenerateQuestionsBodyCategory } from "./generateQuestionsBodyCategory";
import type { GenerateQuestionsBodyDifficulty } from "./generateQuestionsBodyDifficulty";

export interface GenerateQuestionsBody {
  category: GenerateQuestionsBodyCategory;
  difficulty: GenerateQuestionsBodyDifficulty;
  /**
   * @minimum 1
   * @maximum 20
   */
  count?: number;
}