# lib/api-zod/src/generated/types/generatedQuestion.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface GeneratedQuestion {
  id: string;
  question: string;
  correctAnswer: string;
  /**
   * @minItems 3
   * @maxItems 3
   */
  incorrectAnswers: string[];
  category: string;
  difficulty: string;
}