# lib/api-zod/src/generated/types/generatedQuestionsResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { GeneratedQuestion } from "./generatedQuestion";
import type { GeneratedQuestionsResponseSource } from "./generatedQuestionsResponseSource";

export interface GeneratedQuestionsResponse {
  questions: GeneratedQuestion[];
  source: GeneratedQuestionsResponseSource;
}